requirements.txt
django-bootstrap4
request
